+++ benchmark/btb_benchmark/main.py
@@ -13,6 +13,7 @@ from btb.tuning.tuners.base import BaseTuner
 from btb_benchmark.challenges import (
     MATH_CHALLENGES, RandomForestChallenge, SGDChallenge, XGBoostChallenge)
 from btb_benchmark.challenges.challenge import Challenge
+from btb_benchmark.challenges.datasets import get_dataset_names
 from btb_benchmark.results import load_results, write_results
 from btb_benchmark.tuning_functions import get_all_tuning_functions
 from btb_benchmark.tuning_functions.btb import make_btb_tuning_function
@@ -21,7 +22,7 @@ LOGGER = logging.getLogger(__name__)
 ALL_TYPES = ['math', 'xgboost']
 
 
-def get_math_challenge_instance(name):
+def get_math_challenge_instance(name, *args, **kwargs):
     return MATH_CHALLENGES.get(name)()
 
 
@@ -49,6 +50,8 @@ def _evaluate_tuner_on_challenge(name, tuner, challenge, iterations):
             'elapsed': datetime.utcnow() - start,
             'hostname': socket.gethostname()
         }
+        if hasattr(challenge, 'data'):
+            result['rows'] = challenge.data[0].shape[0]
 
     except Exception as ex:
         LOGGER.warn(
@@ -180,6 +183,14 @@ def _as_list(param):
     return [param]
 
 
+def _challenges_as_list(param):
+    """Make sure that param is either ``None`` or a ``list``."""
+    if param is None or isinstance(param, (list, tuple)):
+        return param
+
+    return get_dataset_names(param)
+
+
 def _get_tuners_dict(tuners=None):
     all_tuners = get_all_tuning_functions()
     if tuners is None:
@@ -211,14 +222,15 @@ def _get_all_challenge_names(challenge_types=None):
     if 'math' in challenge_types:
         all_challenge_names += list(MATH_CHALLENGES.keys())
     if any(name in challenge_types for name in ('sdg', 'xgboost', 'random_forest')):
-        all_challenge_names += SGDChallenge.get_available_dataset_names()
+        all_challenge_names += get_dataset_names('all')
 
     return all_challenge_names
 
 
-def _get_challenges_list(challenges=None, challenge_types=None, sample=None):
+def _get_challenges_list(challenges=None, challenge_types=None, sample=None, max_rows=None):
     challenge_types = _as_list(challenge_types) or ALL_TYPES
-    challenges = _as_list(challenges) or _get_all_challenge_names(challenge_types)
+    challenges = _challenges_as_list(challenges) or _get_all_challenge_names(challenge_types)
+
     selected = []
     unknown = []
 
@@ -235,7 +247,9 @@ def _get_challenges_list(challenges=None, challenge_types=None, sample=None):
         else:
             for challenge_type in challenge_types:
                 try:
-                    challenge_instance = CHALLENGE_GETTER[challenge_type](challenge)
+                    challenge_class = CHALLENGE_GETTER[challenge_type]
+                    challenge_instance = challenge_class(challenge, max_rows=max_rows)
+
                     if challenge_instance:
                         known = True
                         selected.append(challenge_instance)
@@ -254,8 +268,8 @@ def _get_challenges_list(challenges=None, challenge_types=None, sample=None):
     return selected
 
 
-def run_benchmark(tuners=None, challenge_types=None, challenges=None,
-                  sample=None, iterations=100, output_path=None, detailed_output=False):
+def run_benchmark(tuners=None, challenge_types=None, challenges=None, sample=None,
+                  iterations=100, max_rows=None, output_path=None, detailed_output=False):
     """Execute the benchmark function and optionally store the result as a ``CSV``.
 
     This function provides a user-friendly interface to interact with the ``benchmark``
@@ -270,14 +284,19 @@ def run_benchmark(tuners=None, challenge_types=None, challenges=None,
         challenge_types (str or list):
             Type or list of types for challenges to be benchmarked, if ``None`` all available
             types will be used.
-        challenges (str, btb_benchmark.challenge.Challenge or list):
-            Challenge name, ``btb_benchmark.challenge.Challenge`` instance or a list with the
-            previously described objects. If ``None`` will use ``challenge_types`` to determine
-            which challenges to use.
+        challenges (str or list):
+            If ``str`` it will be interpreted as ``collection`` of datasets (currently: all or
+            openml100). A list containing: challenge name, ``btb_benchmark.challenge.Challenge``
+            instance or a list with the previously described objects. If ``None`` will use
+            ``challenge_types`` to determine which challenges to use.
         sample (int):
             Run only on a subset of the available datasets of the given size.
         iterations (int):
             Number of tuning iterations to perform per challenge and tuner.
+        max_rows (int):
+            Maximum number of rows to use from each dataset. If ``None``, or if the
+            given number is higher than the number of rows in the dataset, the entire
+            dataset is used. Defaults to ``None``.
         output_path (str):
             If an ``output_path`` is given, the final results will be saved in that location.
         detailed_output (bool):
@@ -292,7 +311,8 @@ def run_benchmark(tuners=None, challenge_types=None, challenges=None,
     challenges = _get_challenges_list(
         challenges=challenges,
         challenge_types=challenge_types,
-        sample=sample
+        sample=sample,
+        max_rows=max_rows
     )
 
     results = benchmark(tuners, challenges, iterations, detailed_output)